Remove the licenseloader.
